Removing unnecessary creation of a SecurityManager. Fixes #176.
Previous code was constructing a SecurityManager and failing to catch
the SecurityException which this could throw.
Thread.currentThread().getStackTrace() is the modern API for looking
at the current stack, and much more likely to be permitted by the
security policy, and there was already code calling this newer method,
but for some reason it was only done as a fallback.

This change rewrites the method with the newer API as the only way.