Add support for more platforms in OSGi testing
The list of supported platforms for OSGi testing isn't intended to be exhaustive, as at some point we stop testing JavaCPP and start testing Maven/OSGi, however it does need to reflect the platforms that JavaCPP is built on. The list has therefore been extended to cover windows, and a wider variety of CPU archetectures on Linux.

Android and iOS have been ignored deliberately as the Maven build tool doesn't natively support them. They could be added in future if this requirement changes

Signed-off-by: Tim Ward <timothyjward@apache.org>