Do not avoid copying array content for const ** arrays. Do that for `* const`s.
This fixes https://github.com/bytedeco/javacpp/issues/776

Before this change, we did not copy array contents back to the argument array
for "const **" as it is regarded as immutable. But it in fact it mutable.
This StackOverflow question describes it well.
https://stackoverflow.com/questions/4949254/const-char-const-versus-const-char

After this change, it will avoid copying if the parameter pointer itself is
const.